Created SDL2-compatible sound interface.
This is a fork from libsndifsdl. There have been some major changes
in the code, so the version number will be raised from 0.7 to 0.8
for this module.

For pratical purposes this should behave exactly like the SDL1-
compatible version. The only relevant changes are that the
"force-8bit-sound" configuration option is gone and that samples
may now use up to 32bits.